refactor: :core:data モジュールについて、data -> core.data に変更した。
