+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/RepositorySearchFragment.kt
@@ -12,7 +12,7 @@ import androidx.fragment.app.viewModels
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.fragment.findNavController
 import dagger.hilt.android.AndroidEntryPoint
-import jp.co.yumemi.android.codecheck.data.model.RepositoryItem
+import jp.co.yumemi.android.codecheck.core.data.model.RepositoryItem
 import jp.co.yumemi.android.codecheck.designsystem.MainTheme
 import jp.co.yumemi.android.codecheck.search.RepositorySearchScreen
 import jp.co.yumemi.android.codecheck.search.RepositorySearchViewModel
