+++ app/src/main/res/navigation/nav_graph.xml
@@ -19,7 +19,7 @@
         android:label="@string/app_name">
         <argument
             android:name="item"
-            app:argType="jp.co.yumemi.android.codecheck.data.model.RepositoryItem" />
+            app:argType="jp.co.yumemi.android.codecheck.core.data.model.RepositoryItem" />
     </fragment>
 
 </navigation>
