+++ core/data/src/main/java/jp/co/yumemi/android/codecheck/core/data/model/DataLoadingState.kt
@@ -1,4 +1,4 @@
-package jp.co.yumemi.android.codecheck.data.model
+package jp.co.yumemi.android.codecheck.core.data.model
 
 /**
  * データの読み込み状態を表す
