+++ core/data/src/test/java/jp/co/yumemi/android/codecheck/core/data/model/ErrorTypeTest.kt
@@ -1,4 +1,4 @@
-package jp.co.yumemi.android.codecheck.data.model
+package jp.co.yumemi.android.codecheck.core.data.model
 
 import java.net.UnknownHostException
 import kotlin.test.Test
