+++ feature/search/src/main/java/jp/co/yumemi/android/codecheck/search/RepositorySearchUiState.kt
@@ -1,7 +1,7 @@
 package jp.co.yumemi.android.codecheck.search
 
-import jp.co.yumemi.android.codecheck.data.model.DataLoadingState
-import jp.co.yumemi.android.codecheck.data.model.RepositoryItem
+import jp.co.yumemi.android.codecheck.core.data.model.DataLoadingState
+import jp.co.yumemi.android.codecheck.core.data.model.RepositoryItem
 
 data class RepositorySearchUiState(
     val dataLoadingState: DataLoadingState = DataLoadingState.Initial,
